requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
Django==3.2.12
gunicorn==20.1.0
importlib-resources==5.4.0
Pillow==8.4.0
pkg-resources==0.0.0
psycopg2-binary==2.9.3
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
zipp==3.6.0
